need to finish login problem

+++ src/Controller/TicketsController.php
@@ -14,6 +14,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 #[Route('/tickets')]
 final class TicketsController extends AbstractController
@@ -21,6 +22,7 @@ final class TicketsController extends AbstractController
     #[Route(name: 'app_tickets_index', methods: ['GET'])]
     public function index(TicketsRepository $ticketsRepository): Response
     {
+
         return $this->render('tickets/index.html.twig', [
             'tickets' => $ticketsRepository->findAll(),
             'stacks' => Stacks::cases(),
@@ -30,9 +32,10 @@ final class TicketsController extends AbstractController
     }
 
     #[Route('/new', name: 'app_tickets_new', methods: ['GET', 'POST'])]
+    #[IsGranted('ROLE_USER')]
     public function new(Request $request, EntityManagerInterface $entityManager, StatusRepository $statusRepo): Response
     {
-        // $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        
         $ticket = new Tickets();
 
         // status Open by defauly
@@ -59,6 +62,7 @@ final class TicketsController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_tickets_show', methods: ['GET'])]
+    #[IsGranted('ROLE_USER')]
     public function show(Tickets $ticket): Response
     {
         return $this->render('tickets/show.html.twig', [
@@ -66,27 +70,29 @@ final class TicketsController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}/edit', name: 'app_tickets_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Tickets $ticket, EntityManagerInterface $entityManager): Response
-    {
-        $form = $this->createForm(TicketsType::class, $ticket);
-        $form->handleRequest($request);
+    // #[Route('/{id}/edit', name: 'app_tickets_edit', methods: ['GET', 'POST'])]
+    // public function edit(Request $request, Tickets $ticket, EntityManagerInterface $entityManager): Response
+    // {
+    //     $form = $this->createForm(TicketsType::class, $ticket);
+    //     $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager->flush();
+    //     if ($form->isSubmitted() && $form->isValid()) {
+    //         $entityManager->flush();
 
-            return $this->redirectToRoute('app_tickets_index', [], Response::HTTP_SEE_OTHER);
-        }
+    //         return $this->redirectToRoute('app_tickets_index', [], Response::HTTP_SEE_OTHER);
+    //     }
 
-        return $this->render('tickets/edit.html.twig', [
-            'ticket' => $ticket,
-            'form' => $form,
-        ]);
-    }
+    //     return $this->render('tickets/edit.html.twig', [
+    //         'ticket' => $ticket,
+    //         'form' => $form,
+    //     ]);
+    // }
 
     #[Route('/{id}', name: 'app_tickets_delete', methods: ['POST'])]
     public function delete(Request $request, Tickets $ticket, EntityManagerInterface $entityManager): Response
     {
+
+        $this->denyAccessUnlessGranted('TICKET_DELETE', $ticket);
         if ($this->isCsrfTokenValid('delete' . $ticket->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($ticket);
             $entityManager->flush();
